pyproject.toml
[tool.poetry]
name = "nous"
version = "0.1.0"
description = ""
authors = ["anthoeknee <pacyheb@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
networkx = "^3.4.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
redis = "^5.2.1"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
google-genai = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
